Light/Dark Mode in the navbar and body (this will changed to context in future)
