+++ src/styles/globals.css
@@ -1,4 +1,3 @@
-
 * {
   box-sizing: border-box;
   padding: 0;
@@ -15,6 +14,8 @@
   height: 100vh;
   padding: 0 170px;
 }
+
+
 
 
 /* Alteração para teste */
